reorganize descriptions into framework setup
